unify check combine
